requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
Pillow==10.3.0
google-generativeai==0.7.2
groq==0.9.0
requests==2.32.3
pydantic==2.7.4
